dynamically set the name of the ldr tool
The ldr tool has been added to the repository as a binary, and named with the
expectation that the user is performing a very specific build using the poky
distro.

Modify the name of the ldr tool to suit the specific build being performed so
it will be found by the u-boot Makefile. This makes the build more flexible by
naming the ldr tool on-the-fly based on the build being performed.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>