remove HOST_PREFIX/CROSS_COMPILE naming for ldr
The upstream U-Boot Makefile contains the line:

	$(CROSS_COMPILE)ldr

for calling the ldr program in order to generate the required boot stream for
booting from an sc5xx's bootrom. This is the exact same prefix string that is
used by gcc's suite of tools. This prefix depends on the target architecture,
distribution, OS, and ABI calling convention. E.g. arm-poky-linux-gnueabi-ldr.

This works when builds are run with these exact assumptions, but if anything
changes, the compile step of u-boot-adi will fail:

	| arm-oe-linux-gnueabi-ldr -T "ADSP-SC589-0.1" -c u-boot.ldr u-boot --bcode=1 --use-vmas
	| /bin/sh: arm-oe-linux-gnueabi-ldr: command not found

Fix this issue in a generic way so that it works for all situations.

Specifically in the ldr-adi-native recipe do not install the ldr utility with
a prefix. Also, in the u-boot-adi recipe, set an environment variable when
building to ignore the $(CROSS_COMPILE) component when looking for the ldr
tool.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>